- fixed call to boot_page_fill_safe(), must be in NRWW section, otherwise SPM will not work - added call to boot_rww_enable() after page write to have access to those code parts of the bootloader that reside in RWW section. - fixed unintended 16 bit arithmetic in checksum_16 function

+++ checksum.c
@@ -70,7 +70,7 @@ uint16_t Checksum_16 (unsigned char * pointer, uint16_t headerlenght, unsigned l
 
 	//Komplementbildung (addiert Long INT_H Word mit Long INT L Word)
 	while (checksum.nWordAcc.DataH)
-		checksum.lLongAcc = checksum.nWordAcc.DataL + checksum.nWordAcc.DataH;
+		checksum.lLongAcc = (uint32_t)checksum.nWordAcc.DataL + (uint32_t)checksum.nWordAcc.DataH;
 	return ~(checksum.nWordAcc.DataL);
 }
